app/profiles/serializers.py
# profiles/serializers.py
from rest_framework import serializers
from .models import UserProfile, Preference

class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = '__all__'

class PreferenceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Preference
        fields = '__all__'
